+++ module.go
@@ -1,12 +1,14 @@
-package core
+package main
 
 import (
 	"github.com/Becklyn/go-fx-core/env"
-	"github.com/Becklyn/go-fx-core/fiber"
+	f "github.com/Becklyn/go-fx-core/fiber"
 	"github.com/Becklyn/go-fx-core/health"
 	"github.com/Becklyn/go-fx-core/logrus"
 	"github.com/Becklyn/go-fx-core/metrics"
+	"github.com/Becklyn/go-fx-core/middleware"
 	"github.com/Becklyn/go-fx-core/readyness"
+	"github.com/gofiber/fiber/v2"
 	"go.uber.org/fx"
 )
 
@@ -16,5 +18,17 @@ var Module = fx.Options(
 	metrics.Module,
 	readyness.Module,
 	health.Module,
-	fiber.Module,
+	middleware.Module,
+	f.Module,
 )
+
+func main() {
+	fx.New(
+		Module,
+		fx.Invoke(func(app *fiber.App) {
+			app.Get("/", func(c *fiber.Ctx) error {
+				return c.SendString("Hello, World!")
+			})
+		}),
+	).Run()
+}